+++ Tests/Standard.AspNetCore.TestApplication/Standard.AspNetCore.TestApplication.Application/Interfaces/IQueryStringNamesService.cs
@@ -1,0 +1,15 @@
+using System;
+using System.Threading;
+using System.Threading.Tasks;
+using Intent.RoslynWeaver.Attributes;
+
+[assembly: DefaultIntentManaged(Mode.Fully)]
+[assembly: IntentTemplate("Intent.Application.Contracts.ServiceContract", Version = "1.0")]
+
+namespace Standard.AspNetCore.TestApplication.Application.Interfaces
+{
+    public interface IQueryStringNamesService : IDisposable
+    {
+        Task Operation(string par1, CancellationToken cancellationToken = default);
+    }
+}