+++ Modules/Intent.Modules.AspNetCore.Controllers.Dispatch.MediatR/release-notes.md
@@ -1,3 +1,7 @@
+### Version 5.4.1
+
+- Improvement: Will now respect query string parameter names as introduced in [`Intent.Metadata.WebApi` version `4.3.2`](https://github.com/IntentArchitect/Intent.Modules/blob/development/Modules/Intent.Modules.Metadata.WebApi/release-notes.md#version-432).
+
 ### Version 5.4.0
 
 - Improvement: When Controller parameters match with Update Command fields, we've made consuming these endpoints easier. Now, you don't need to populate the fields on the Command that are already populated via a Route parameter.